+++ .eslintrc.js
@@ -1,15 +1,21 @@
 module.exports = {
-  "extends": "standard",
-  "plugins": [
-    "standard", "node", "import", "promise", "react", "emotion"
-  ],
-  "rules": {
-    "space-before-function-paren": ["error", "never"],
-    "semi": ["error", "always"],
-    "quotes": ["error", "double"],
-    "keyword-spacing": ["error", {"before": true, "after": false, "overrides": {"else": {"after": true}, "from": {"after": true}, "import": {"after": true}, "case": {"after": true}}}],
-    "space-unary-ops": [2, {"words": true, "nonwords": false, "overrides": {"typeof": false}}],
+  env: {
+    browser: true,
+    es6: true,
+    node: true,
+  },
+  extends: ["standard", "prettier/standard", "plugin:prettier/recommended"],
+  parser: "@babel/eslint-parser",
+  plugins: ["@emotion", "react", "react-hooks", "standard"],
+  rules: {
+    "comma-dangle": ["error", "only-multiline"],
+    "no-unused-vars": [
+      "error",
+      { vars: "all", varsIgnorePattern: "React|^_$", args: "none", ignoreRestSiblings: true },
+    ],
+    "react-hooks/exhaustive-deps": "warn",
+    "react-hooks/rules-of-hooks": "error",
     "react/jsx-uses-vars": [2],
-    "emotion/syntax-preference": [2, "string"]
-  }
+    semi: ["error", "always"],
+  },
 };